requirements.txt
streamlit
pydantic-ai
python-dotenv
nest_asyncio 